Tested SimpleWallet cond_set and cond_add operations

+++ src/secureModule/AddRequest.java
@@ -3,10 +3,10 @@ package secureModule;
 public class AddRequest {
 
 	public long opi;
-	public int amount;
+	public long amount;
 	public String cipheredKey;
 	
-	public AddRequest(long opi, int amount, String cipheredKey) {
+	public AddRequest(long opi, long amount, String cipheredKey) {
 		this.opi = opi;
 		this.amount = amount;
 		this.cipheredKey = cipheredKey;
